updated system prompt and moved everythign to pr_reviewer
